coordinator: organise packages
