+++ coordinator/app/src/main/kotlin/net/consensys/zkevm/coordinator/app/BlockchainClientHelper.kt
@@ -3,6 +3,7 @@ package net.consensys.zkevm.coordinator.app
 import io.vertx.core.Vertx
 import io.vertx.core.http.HttpVersion
 import io.vertx.ext.web.client.WebClientOptions
+import linea.web3j.SmartContractErrors
 import net.consensys.linea.contract.AsyncFriendlyTransactionManager
 import net.consensys.linea.contract.EIP1559GasProvider
 import net.consensys.linea.contract.L2MessageService
@@ -13,7 +14,6 @@ import net.consensys.linea.ethereum.gaspricing.FeesCalculator
 import net.consensys.linea.ethereum.gaspricing.FeesFetcher
 import net.consensys.linea.ethereum.gaspricing.WMAGasProvider
 import net.consensys.linea.httprest.client.VertxHttpRestClient
-import net.consensys.linea.web3j.SmartContractErrors
 import net.consensys.zkevm.coordinator.app.config.L1Config
 import net.consensys.zkevm.coordinator.app.config.L2Config
 import net.consensys.zkevm.coordinator.app.config.SignerConfig
