+++ coordinator/app/src/main/kotlin/net/consensys/zkevm/coordinator/app/L1toL2MessageAnchoringApp.kt
@@ -1,11 +1,11 @@
 package net.consensys.zkevm.coordinator.app
 
 import io.vertx.core.Vertx
+import linea.web3j.SmartContractErrors
 import net.consensys.linea.contract.AsyncFriendlyTransactionManager
 import net.consensys.linea.contract.EIP1559GasProvider
 import net.consensys.linea.contract.L2MessageService
 import net.consensys.linea.contract.l1.Web3JLineaRollupSmartContractClient
-import net.consensys.linea.web3j.SmartContractErrors
 import net.consensys.zkevm.LongRunningService
 import net.consensys.zkevm.coordinator.app.config.L1Config
 import net.consensys.zkevm.coordinator.app.config.L2Config
