+++ coordinator/app/src/main/kotlin/net/consensys/zkevm/coordinator/blockcreation/GethCliqueSafeBlockProvider.kt
@@ -1,8 +1,8 @@
 package net.consensys.zkevm.coordinator.blockcreation
 
-import build.linea.web3j.domain.toWeb3j
 import linea.domain.Block
 import linea.domain.BlockParameter.Companion.toBlockParameter
+import linea.web3j.domain.toWeb3j
 import linea.web3j.toDomain
 import net.consensys.linea.async.toSafeFuture
 import net.consensys.zkevm.ethereum.coordination.blockcreation.SafeBlockProvider
