+++ coordinator/clients/smart-contract-client/src/main/kotlin/net/consensys/linea/contract/EIP1559GasProvider.kt
@@ -2,10 +2,10 @@ package net.consensys.linea.contract
 
 import linea.kotlin.toBigInteger
 import linea.kotlin.toIntervalString
-import net.consensys.linea.web3j.AtomicContractEIP1559GasProvider
-import net.consensys.linea.web3j.EIP1559GasFees
-import net.consensys.linea.web3j.blocksRange
-import net.consensys.linea.web3j.toLineaDomain
+import linea.web3j.domain.blocksRange
+import linea.web3j.domain.toLineaDomain
+import linea.web3j.gas.AtomicContractEIP1559GasProvider
+import linea.web3j.gas.EIP1559GasFees
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 import org.web3j.protocol.Web3j
