+++ coordinator/clients/smart-contract-client/src/main/kotlin/net/consensys/linea/contract/LineaRollupAsyncFriendly.kt
@@ -3,9 +3,9 @@ package net.consensys.linea.contract
 import build.linea.contract.LineaRollupV6
 import linea.kotlin.toBigInteger
 import linea.kotlin.toULong
-import net.consensys.linea.web3j.AtomicContractEIP1559GasProvider
-import net.consensys.linea.web3j.EIP1559GasFees
-import net.consensys.linea.web3j.SmartContractErrors
+import linea.web3j.SmartContractErrors
+import linea.web3j.gas.AtomicContractEIP1559GasProvider
+import linea.web3j.gas.EIP1559GasFees
 import org.web3j.abi.FunctionEncoder
 import org.web3j.abi.datatypes.Function
 import org.web3j.crypto.Credentials
