+++ coordinator/clients/smart-contract-client/src/main/kotlin/net/consensys/linea/contract/Web3JContractAsyncHelper.kt
@@ -3,15 +3,15 @@ package net.consensys.linea.contract
 import linea.kotlin.toBigInteger
 import linea.kotlin.toGWei
 import linea.kotlin.toULong
+import linea.web3j.SmartContractErrors
+import linea.web3j.domain.Eip4844Transaction
+import linea.web3j.gas.AtomicContractEIP1559GasProvider
+import linea.web3j.gas.EIP1559GasFees
+import linea.web3j.gas.EIP4844GasFees
+import linea.web3j.gas.EIP4844GasProvider
+import linea.web3j.getRevertReason
+import linea.web3j.informativeEthCall
 import net.consensys.linea.async.toSafeFuture
-import net.consensys.linea.web3j.AtomicContractEIP1559GasProvider
-import net.consensys.linea.web3j.EIP1559GasFees
-import net.consensys.linea.web3j.EIP4844GasFees
-import net.consensys.linea.web3j.EIP4844GasProvider
-import net.consensys.linea.web3j.Eip4844Transaction
-import net.consensys.linea.web3j.SmartContractErrors
-import net.consensys.linea.web3j.getRevertReason
-import net.consensys.linea.web3j.informativeEthCall
 import net.consensys.zkevm.domain.BlobRecord
 import net.consensys.zkevm.ethereum.gaspricing.GasPriceCaps
 import org.apache.logging.log4j.LogManager
