+++ coordinator/clients/smart-contract-client/src/main/kotlin/net/consensys/linea/contract/Web3JL2MessageServiceLogsClient.kt
@@ -1,7 +1,7 @@
 package net.consensys.linea.contract
 
-import build.linea.web3j.Web3JLogsClient
 import linea.kotlin.toULong
+import linea.web3j.Web3JLogsClient
 import net.consensys.zkevm.coordinator.clients.L2MessageServiceLogsClient
 import net.consensys.zkevm.domain.BridgeLogsData
 import net.consensys.zkevm.domain.L2RollingHashUpdatedEvent
