+++ coordinator/clients/smart-contract-client/src/main/kotlin/net/consensys/linea/contract/l1/Web3JLineaRollupSmartContractClient.kt
@@ -3,10 +3,10 @@ package net.consensys.linea.contract.l1
 import build.linea.contract.LineaRollupV6
 import build.linea.contract.l1.Web3JLineaRollupSmartContractClientReadOnly
 import linea.kotlin.toULong
+import linea.web3j.SmartContractErrors
 import net.consensys.linea.contract.AsyncFriendlyTransactionManager
 import net.consensys.linea.contract.Web3JContractAsyncHelper
 import net.consensys.linea.contract.throwExceptionIfJsonRpcErrorReturned
-import net.consensys.linea.web3j.SmartContractErrors
 import net.consensys.zkevm.coordinator.clients.smartcontract.BlockAndNonce
 import net.consensys.zkevm.coordinator.clients.smartcontract.LineaRollupSmartContractClient
 import net.consensys.zkevm.domain.BlobRecord
