+++ coordinator/ethereum/gas-pricing/dynamic-cap/src/main/kotlin/net/consensys/linea/ethereum/gaspricing/dynamiccap/GasPriceCapFeeHistoryFetcherImpl.kt
@@ -1,7 +1,7 @@
 package net.consensys.linea.ethereum.gaspricing.dynamiccap
 
 import linea.domain.FeeHistory
-import net.consensys.linea.web3j.Web3jBlobExtended
+import linea.web3j.Web3jBlobExtended
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 import org.web3j.protocol.core.DefaultBlockParameter
