+++ coordinator/ethereum/gas-pricing/static-cap/src/main/kotlin/net/consensys/linea/ethereum/gaspricing/staticcap/FeeHistoryFetcherImpl.kt
@@ -2,8 +2,8 @@ package net.consensys.linea.ethereum.gaspricing.staticcap
 
 import linea.domain.FeeHistory
 import linea.kotlin.toIntervalString
+import linea.web3j.Web3jBlobExtended
 import net.consensys.linea.ethereum.gaspricing.FeesFetcher
-import net.consensys.linea.web3j.Web3jBlobExtended
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 import org.web3j.protocol.Web3j
