+++ coordinator/ethereum/test-utils/src/main/kotlin/net/consensys/zkevm/ethereum/ContractsManager.kt
@@ -4,6 +4,7 @@ import com.sksamuel.hoplite.ConfigLoaderBuilder
 import com.sksamuel.hoplite.addFileSource
 import linea.contract.l1.LineaContractVersion
 import linea.kotlin.gwei
+import linea.web3j.SmartContractErrors
 import net.consensys.linea.contract.AsyncFriendlyTransactionManager
 import net.consensys.linea.contract.EIP1559GasProvider
 import net.consensys.linea.contract.LineaRollupAsyncFriendly
@@ -11,7 +12,6 @@ import net.consensys.linea.contract.StaticGasProvider
 import net.consensys.linea.contract.l1.Web3JLineaRollupSmartContractClient
 import net.consensys.linea.contract.l2.L2MessageServiceGasLimitEstimate
 import net.consensys.linea.testing.filesystem.findPathTo
-import net.consensys.linea.web3j.SmartContractErrors
 import net.consensys.zkevm.coordinator.clients.smartcontract.LineaRollupSmartContractClient
 import org.slf4j.LoggerFactory
 import org.web3j.protocol.Web3j
