+++ finalized-tag-updater/src/main/kotlin/net/consensys/linea/LineaL1FinalizationUpdaterService.kt
@@ -4,7 +4,7 @@ import build.linea.contract.l1.Web3JLineaRollupSmartContractClientReadOnly
 import io.vertx.core.Vertx
 import linea.consensus.EngineBlockTagUpdater
 import linea.domain.BlockParameter
-import net.consensys.linea.web3j.okHttpClientBuilder
+import linea.web3j.okhttp.okHttpClientBuilder
 import net.consensys.zkevm.LongRunningService
 import net.consensys.zkevm.ethereum.finalization.FinalizationUpdatePoller
 import net.consensys.zkevm.ethereum.finalization.FinalizationUpdatePollerConfig
