+++ jvm-libs/linea/web3j-extensions/src/main/kotlin/linea/web3j/Blob.kt
@@ -1,4 +1,4 @@
-package net.consensys.linea.web3j
+package linea.web3j
 
 import linea.domain.Constants
