+++ jvm-libs/linea/web3j-extensions/src/main/kotlin/linea/web3j/EthCallWithInformativeReverts.kt
@@ -1,4 +1,4 @@
-package net.consensys.linea.web3j
+package linea.web3j
 
 import org.web3j.protocol.Web3j
 import org.web3j.protocol.core.DefaultBlockParameterName
