+++ jvm-libs/linea/web3j-extensions/src/main/kotlin/linea/web3j/EthFeeHistoryBlobExtended.kt
@@ -1,4 +1,4 @@
-package net.consensys.linea.web3j
+package linea.web3j
 
 import com.fasterxml.jackson.core.JsonParser
 import com.fasterxml.jackson.core.JsonToken
