+++ jvm-libs/linea/web3j-extensions/src/main/kotlin/linea/web3j/ExtendedWeb3J.kt
@@ -1,9 +1,8 @@
-package net.consensys.linea.web3j
+package linea.web3j
 
-import build.linea.web3j.domain.toWeb3j
 import linea.domain.Block
 import linea.domain.BlockParameter
-import linea.web3j.toDomain
+import linea.web3j.domain.toWeb3j
 import net.consensys.linea.async.toSafeFuture
 import org.web3j.protocol.Web3j
 import org.web3j.protocol.core.DefaultBlockParameter
