+++ jvm-libs/linea/web3j-extensions/src/main/kotlin/linea/web3j/Web3JFactory.kt
@@ -1,6 +1,6 @@
 package linea.web3j
 
-import net.consensys.linea.web3j.okHttpClientBuilder
+import linea.web3j.okhttp.okHttpClientBuilder
 import org.apache.logging.log4j.Level
 import org.apache.logging.log4j.Logger
 import org.web3j.protocol.Web3j
