+++ jvm-libs/linea/web3j-extensions/src/main/kotlin/linea/web3j/Web3JLogsClient.kt
@@ -1,4 +1,4 @@
-package build.linea.web3j
+package linea.web3j
 
 import io.vertx.core.Vertx
 import net.consensys.linea.async.AsyncRetryer
