+++ jvm-libs/linea/web3j-extensions/src/main/kotlin/linea/web3j/domain/Eip4844EthCall.kt
@@ -1,4 +1,4 @@
-package net.consensys.linea.web3j
+package linea.web3j.domain
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.core.JsonGenerator
