+++ jvm-libs/linea/web3j-extensions/src/main/kotlin/linea/web3j/domain/FeeHistoryExtensions.kt
@@ -1,4 +1,4 @@
-package net.consensys.linea.web3j
+package linea.web3j.domain
 
 import linea.domain.FeeHistory
 import linea.kotlin.toULong
