+++ jvm-libs/linea/web3j-extensions/src/main/kotlin/linea/web3j/gas/EIP4844GasProvider.kt
@@ -1,4 +1,4 @@
-package net.consensys.linea.web3j
+package linea.web3j.gas
 
 import org.web3j.tx.gas.ContractEIP1559GasProvider
