+++ jvm-libs/linea/web3j-extensions/src/main/kotlin/linea/web3j/okhttp/OkHttpClientLogging.kt
@@ -1,4 +1,4 @@
-package net.consensys.linea.web3j
+package linea.web3j.okhttp
 
 import net.consensys.linea.logging.JsonRpcRequestResponseLogger
 import net.consensys.linea.logging.MinimalInLineJsonRpcLogger
