+++ transaction-decoder-tool/src/test/kotlin/linea/test/BlocksFetcher.kt
@@ -1,9 +1,9 @@
 package linea.test
 
-import build.linea.web3j.domain.toWeb3j
 import io.vertx.core.Vertx
 import linea.domain.Block
 import linea.domain.BlockParameter.Companion.toBlockParameter
+import linea.web3j.domain.toWeb3j
 import linea.web3j.toDomain
 import net.consensys.linea.async.AsyncRetryer
 import net.consensys.linea.async.toSafeFuture
